Converted size field (IAModeler class) to BigDecimal (#24)

+++ src/main/java/org/eclipse/opensmartclide/dbapi/model/IAModeler.java
@@ -11,6 +11,7 @@ import org.springframework.format.annotation.DateTimeFormat;
 
 import javax.validation.constraints.NotEmpty;
 import javax.validation.constraints.NotNull;
+import java.math.BigDecimal;
 import java.util.ArrayList;
 import java.util.Date;
 import java.util.List;
@@ -40,7 +41,7 @@ public class IAModeler {
 
     private String target;
 
-    private Float size;
+    private BigDecimal size;
 
     @Field
     private List<String> labels = new ArrayList<>();
